+++ tests/test_run.py
@@ -1,3 +1,5 @@
+import os
+
 from bocadillo import App
 from tests.utils import env
 
@@ -36,13 +38,32 @@ def test_port_is_8000_by_default(app: App):
     app.run(_run=run)
 
 
-def test_if_debug_then_debug_mode_activated(app: App):
+def test_debug_mode_disabled_by_default(app: App):
+    def run(app, **kwargs):
+        pass
+
+    assert not app.debug
+    app.run(_run=run)
+    assert not app.debug
+
+
+def test_if_debug_then_debug_mode_enabled(app: App):
     def run(app, **kwargs):
         pass
 
     assert not app.debug
     app.run(debug=True, _run=run)
     assert app.debug
+
+
+def test_if_debug_env_var_set_then_debug_mode_enabled(app: App):
+    os.environ["BOCADILLO_DEBUG"] = "1"
+    try:
+        assert not app.debug
+        app.run(_run=lambda app, **kwargs: None)
+        assert app.debug
+    finally:
+        os.environ.pop("BOCADILLO_DEBUG")
 
 
 def test_if_debug_then_app_given_as_import_string(app: App):